code for monitoring keyboard compartments (https://blogs.msdn.microsoft.com/tsfaware/2007/05/30/whats-a-keyboard/)

+++ VietTypeATL/TextService.cpp
@@ -93,7 +93,7 @@ STDMETHODIMP VietType::TextService::ActivateEx(ITfThreadMgr * ptim, TfClientId t
     hr = _engineController->Initialize(_engine, ptim, tid);
     HRESULT_CHECK_RETURN(hr, L"%s", L"_engineController->Initialize failed");
 
-    int enabled;
+    long enabled;
     hr = _engineController->IsUserEnabled(&enabled);
     HRESULT_CHECK_RETURN(hr, L"%s", L"_engineController->IsUserEnabled failed");
     hr = _engineController->WriteUserEnabled(enabled);
